requirements.txt
# Web stack
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.2.0

# Testing
httpx>=0.27.0
trio>=0.25.0

# Desktop compatibility
PyQt5==5.15.9
qtawesome>=1.4.0
schedule>=1.2.0

# Directory and HTTP
ldap3>=2.9.1
requests>=2.28.0

# Packaging
pyinstaller>=6.0.0
Pillow>=10.0.0
